:sparkles: Show threat info and stats on repositories table
