+++ components/repositories/AccountView.tsx
@@ -61,6 +61,7 @@ import { DidHistory } from './DidHistory'
 import { InviteCodeGenerationStatus } from './InviteCodeGenerationStatus'
 import { MuteReporting } from './MuteReporting'
 import { ProfileAvatar } from './ProfileAvatar'
+import { obscureIp, parseThreatSigs } from './helpers'
 
 enum Views {
   Details,
@@ -953,52 +954,4 @@ const EmailView = (props: ComponentProps<typeof EmailComposer>) => {
       <EmailComposer {...props} />
     </div>
   )
-}
-
-function obscureIp(ip: string) {
-  const parts = ip.split('.')
-  if (parts.length !== 4) return '***.***.***.***'
-  return `${parts[0]}.${parts[1]}.***.***`
-}
-
-function parseThreatSigs(sigs?: ComAtprotoAdminDefs.ThreatSignature[]) {
-  const hcapDetail: ComAtprotoAdminDefs.ThreatSignature[] = []
-  let registrationIp,
-    lastSigninIp,
-    lastSigninTime,
-    lastSigninCountry,
-    ipCountry: string | undefined
-
-  if (sigs) {
-    for (const sig of sigs) {
-      switch (sig.property) {
-        case 'registrationIp':
-          registrationIp = sig.value
-          break
-        case 'lastSigninIp':
-          lastSigninIp = sig.value
-          break
-        case 'lastSigninTime':
-          lastSigninTime = sig.value
-          break
-        case 'lastSigninCountry':
-          lastSigninCountry = sig.value
-          break
-        case 'ipCountry':
-          ipCountry = sig.value
-          break
-        default:
-          hcapDetail.push(sig)
-      }
-    }
-  }
-
-  return {
-    registrationIp,
-    lastSigninIp,
-    lastSigninTime,
-    lastSigninCountry,
-    ipCountry,
-    hcapDetail,
-  }
 }
